Reorganize some types exports/imports, fewer any-type
