+++ x-pack/legacy/plugins/reporting/export_types/common/layouts/preserve_layout.ts
@@ -4,9 +4,8 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import path from 'path';
-import { Size } from '../../../types';
 import { LayoutTypes } from '../constants';
-import { Layout, LayoutSelectorDictionary, PageSizeParams } from './layout';
+import { Layout, LayoutSelectorDictionary, PageSizeParams, Size } from './layout';
 
 // We use a zoom of two to bump up the resolution of the screenshot a bit.
 const ZOOM: number = 2;
